requirements.txt
PyQt5==5.15.9
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0 